Little change

+++ Tourest/Models/Booking.cs
@@ -7,7 +7,7 @@
 		public DateTime DepartureDate { get; set; } // Dùng DateTime, cấu hình thành DATE trong Fluent API
 		public int NumberOfAdults { get; set; }
 		public int NumberOfChildren { get; set; }
-		public decimal TotalPrice { get; set; }
+		public int TotalPrice { get; set; }
 		public string Status { get; set; } = string.Empty;
 		public string? PickupPoint { get; set; }
 		public int CustomerID { get; set; } // FK property
@@ -15,7 +15,7 @@
 		public int? TourGroupID { get; set; } // FK property
 		public int? PaymentID { get; set; } // FK property
 		public DateTime? CancellationDate { get; set; }
-		public decimal? RefundAmount { get; set; }
+		public int? RefundAmount { get; set; }
 
 		// Navigation Properties
 		public virtual User Customer { get; set; } = null!;
